Move getRenderedTab to its own component
